BackEnd/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pillow
moviepy
imageio
imageio-ffmpeg
python-pptx
groq
gtts